pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "charms.proxylib"
dynamic = ["version"]
description = "Charm library for juju-*-proxy model config"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache" }
authors = [
  { name= "Adam Dyess", email = "adam.dyess@canonical.com" },
  { name="Mateo Florido", email="mateo.florido@canonical.com" }
]
dependencies = []
keywords = [
    "juju",
    "charming",
    "operators",
    "systemd",
    "web-proxy",
    "validation"
]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]

[project.optional-dependencies]
lightkube = [
    "lightkube"
]

[project.urls]
Repository = "https://github.com/canonical/charms.proxylib"
BugTracker = "https://github.com/canonical/charms.proxylib/issues"

[tool.ruff]
line-length = 99
extend-exclude = ["__pycache__", "*.egg_info"]

[tool.mypy]
explicit_package_bases = true
namespace_packages = true

[tool.setuptools]
zip-safe = true

[tool.setuptools.packages.find]
namespaces = true

[dependency-groups]
format = ["ruff"]
unit = [
    "lightkube",
    "pytest-cov"
]
lint = [
    "lightkube",
    "mypy",
    "ruff",
    "types-PyYAML",
    "types-backports"
]
